Revert "Lockscreen charging info: show decimal in battery temperature"
This reverts commit 77b344f7f56a91332ed90e6c80a83559a060c937.

+++ packages/SettingsLib/src/com/android/settingslib/fuelgauge/BatteryStatus.java
@@ -64,7 +64,7 @@ public class BatteryStatus {
     public final int maxChargingCurrent;
     public final int maxChargingVoltage;
     public final int maxChargingWattage;
-    public final float temperature;
+    public final int temperature;
     public final boolean present;
 
     public final boolean dashChargeStatus;
@@ -75,7 +75,7 @@ public class BatteryStatus {
 
     public BatteryStatus(int status, int level, int plugged, int health,
             int maxChargingCurrent, int maxChargingVoltage, int maxChargingWattage,
-            float temperature, boolean dashChargeStatus, boolean warpChargeStatus,
+            int temperature, boolean dashChargeStatus, boolean warpChargeStatus,
             boolean voocChargeStatus, boolean miturboChargerStatus, boolean turboPowerStatus, boolean present) {
         this.status = status;
         this.level = level;
